Add "actual" UT for streaming query with Kafka source and sink
